% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{availableDistances}
\alias{availableDistances}
\title{List of available distances}
\usage{
availableDistances()
}
\value{
character vector containing the codes of the available distances
}
\description{
`availableDistance` returns a list of the available codes representing
  distance functions that can be used in the methods of this package.
}
