% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance}
\usage{
distance(data, distance = "euc")
}
\arguments{
\item{distance}{the three-letters name of the distance function chosen
to calculate the distance between the objects of the dataset. Codes:
- Manhattan distance (man):
- Euclidean distance (euc):
- Chebyshev distance (che):}
}
\value{
A matrix containing the distance between all the objects of the
dataset calculating with the chosen distance function.
}
\description{
Distance
}
\examples{

}
