% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distanceBetween}
\alias{distanceBetween}
\title{Calculates the distance between two vector of the same length using the
chosen distance function.}
\usage{
distanceBetween(x, y, distance = "euc")
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{distance}{distance function used to calculate the distance between the
vectors.}
}
\value{
A positive number that is the distance between the two vectors
}
\description{
Calculates the distance between two vector of the same length using the
chosen distance function.
}
\examples{
# Uses the default distance function: Euclidean distance
distanceBetween(c(1, 2, 3), c(3, 2, 1))

}
