% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{getCodeDistance}
\alias{getCodeDistance}
\title{Get the code of a distance function}
\usage{
getCodeDistance(distanceName)
}
\arguments{
\item{distanceName}{Identifier of the distance function using a string of lower case letters of three characters}
}
\value{

}
\description{
Get the code of a distance function
}
\keyword{internal}
