% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDistance}
\alias{plotDistance}
\title{Plot the distance between two functions}
\usage{
plotDistance(index1, index2, df, distance)
}
\arguments{
\item{index1}{Line from}

\item{index2}{Line to}

\item{df}{Dataframe}

\item{distance}{Distance}
}
\value{

}
\description{
Plot the distance between two functions
}
