% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset}
\format{
A tibble with 15 objects and 2 features + real cluster:
\describe{
  \item{V1}{Numeric feature}
  \item{V2}{Numeric feature}
  \item{V2}{Cluster of the object}
}
}
\usage{
toy_data
}
\description{
Toy dataset
}
\keyword{datasets}
