% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data1}
\alias{toy_data1}
\title{Toy dataset}
\format{
A tibble with 5 objects and 2 features:
\describe{
  \item{V1}{Numeric feature}
  \item{V2}{Numeric feature}
}
}
\usage{
toy_data1
}
\description{
Toy dataset
}
\keyword{datasets}
