% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data3}
\alias{toy_data3}
\title{Toy dataset}
\format{
A tibble with 5 objects and 4 features:
\describe{
  \item{V1}{Numeric feature}
  \item{V2}{Numeric feature}
  \item{V3}{Numeric feature}
  \item{V4}{Numeric feature}
}
}
\usage{
toy_data3
}
\description{
Toy dataset
}
\keyword{datasets}
