% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDistance}
\alias{plotDistance}
\title{Plot the distance between two functions}
\usage{
plotDistance(index1, index2, df, distance)
}
\arguments{
\item{distance}{}
}
\value{

}
\description{
Plot the distance between two functions
}
